requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.2.0
backports.functools-lru-cache==1.5
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crayons==0.3.0
Faker==2.0.2
future==0.18.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.1
pytest-rerunfailures==7.0
python-dateutil==2.8.0
requests==2.22.0
selene==1.0.0a16
selenium==3.14.0
six==1.12.0
text-unidecode==1.3
urllib3==1.25.6
wcwidth==0.1.7
webdriver-manager==1.8.2
zipp==0.6.0
